--- src/main/java/InvalidCommandDukeException.java
@@ -1,5 +1,0 @@
-public class InvalidCommandDukeException extends DukeException{
-    public InvalidCommandDukeException(String message) {
-        super(message);
-    }
-}
