--- src/main/java/MissingDescriptionException.java
@@ -1,5 +1,0 @@
-public class MissingDescriptionException extends DukeException{
-    public MissingDescriptionException(String message) {
-        super(message);
-    }
-}
