+++ src/main/java/duke/Query.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Query {
     protected String input;
