+++ src/main/java/duke/QueryType.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum QueryType {
     list,
     mark,
