+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class AddCommand extends Command {
     private Task task;
