+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class ExitCommand extends Command {
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
