+++ src/main/java/duke/command/MarkCommand.java
@@ -1,20 +1,21 @@
-public class DeleteCommand extends Command {
-    private final int index;
+package duke.command;
 
-    public DeleteCommand(int index) {
+import duke.*;
+import duke.task.Task;
+import duke.task.TaskList;
+
+public class MarkCommand extends Command {
+    private final int index;
+    public MarkCommand(int index) {
         this.index = index;
     }
-
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        ui.showMessage("Noted, I've removed this task: ");
+        ui.showMessage("Nice! I've marked this task as done: ");
         Task cur = taskList.tasks.get(index - 1);
-        taskList.deleteTask(index);
+        taskList.markTask(index);
         ui.showMessage("[" + cur.getStatusIcon() + "] " + cur.description);
-        ui.showMessage("Now you have " + taskList.tasks.size() + " tasks in the list.");
     }
-
     public boolean isExit() {
         return false;
     }
 }
-
