+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.*;
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class UnmarkCommand extends Command {
     private final int index;
