+++ src/main/java/duke/task/Deadline.java
@@ -1,7 +1,10 @@
-import java.time.DateTimeException;
+package duke.task;
+
+import duke.DukeException;
+import duke.Parser;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 import java.time.format.FormatStyle;
 
 public class Deadline extends Task {
