+++ src/main/java/duke/task/Todo.java
@@ -1,15 +1,19 @@
+package duke.task;
+
+import duke.DukeException;
+
 public class Todo extends Task {
 
     public Todo(String input, boolean isDone) {
         super(input, isDone);
         this.symbol = 'T';
     }
-    public Todo(String input) throws MissingDescriptionException {
+    public Todo(String input) throws DukeException {
         super(input);
         this.symbol = 'T';
         String[] inputArr = input.split(" ", 2);
         if (inputArr.length == 1 || inputArr[1].isBlank()) {
-            throw new MissingDescriptionException("Sorry, the description of a todo cannot be empty!");
+            throw new DukeException("Sorry, the description of a todo cannot be empty!");
         }
         this.description = inputArr[1];
     }
